requirements.txt
together
flask
python-dotenv
